+++ linux/litertlm_binaries.cmake
@@ -1,4 +1,4 @@
 set(LITERTLM_X86_64_URL
-    "https://files.pythonhosted.org/packages/7a/22/9f40a16fa74b0410658ec1d7372fd52b5a6a6b9ef5864845632538a756d2/litert_lm_api-0.14.0-py3-none-manylinux_2_27_x86_64.whl")
+    "https://files.pythonhosted.org/packages/e3/e3/eb08521ef440ca7967dc8cb057fc5701bc98adcd5cfb78098da8360ade41/litert_lm_api-0.15.0-py3-none-manylinux_2_27_x86_64.whl")
 set(LITERTLM_ARM64_URL
-    "https://files.pythonhosted.org/packages/f8/41/13da5ec4af33394c8bd1138f0ba87776b5fd6e0ba812493ccde44bbc05c/litert_lm_api-0.14.0-py3-none-manylinux_2_27_aarch64.whl")
+    "https://files.pythonhosted.org/packages/60/6d/d7dbeee183fb96be60dc965e0675c4f54e68ae1662da442189332729be58/litert_lm_api-0.15.0-py3-none-manylinux_2_27_aarch64.whl")