+++ windows/litertlm_binaries.cmake
@@ -1,2 +1,2 @@
 set(LITERTLM_X86_64_URL
-    "https://files.pythonhosted.org/packages/d9/ef/36831a18c29b5a8f8283578c77fdd0c823dc8851b05f7a12689ac603c599/litert_lm_api-0.14.0-py3-none-win_amd64.whl")
+    "https://files.pythonhosted.org/packages/c8/41/d70672921bd97c50679845b825dd02bfec875faece2e1529102502b7e6ba/litert_lm_api-0.15.0-py3-none-win_amd64.whl")